correcting in validate options
