+++ models/contacts/contactOperations/validaveOptions.js
@@ -1,3 +1,5 @@
+const Joi = require('joi');
+
 const options = {
   name: {
     pattern: /^[a-zA-Zа-яА-Я]+(([' -][a-zA-Zа-яА-Я ])?[a-zA-Zа-яА-Я]*)*$/,
@@ -8,4 +10,18 @@ const options = {
   },
 };
 
-module.exports = options;
+const { name, number } = options;
+
+const validateAddContact = Joi.object({
+  name: Joi.string().min(3).max(30).pattern(name.pattern, 'name').required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().pattern(number.pattern, 'phone').required(),
+});
+
+const validateUpdateContact = Joi.object({
+  name: Joi.string().min(3).max(30).pattern(name.pattern, 'name'),
+  email: Joi.string().email(),
+  phone: Joi.string().pattern(number.pattern, 'phone'),
+});
+
+module.exports = { validateAddContact, validateUpdateContact };
