+++ models/contacts/index.js
@@ -4,6 +4,7 @@ const removeContact = require('./contactOperations/removeContact');
 const updateContacts = require('./contactOperations/updateContacts');
 const addContact = require('./contactOperations/addContact');
 const updateContactById = require('./contactOperations/updateContactById');
+
 module.exports = {
   listContacts,
   getContactById,
